Allow --config for all commands
