+++ pkg/devspace/config/loader/get.go
@@ -1,7 +1,6 @@
 package loader
 
 import (
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -15,8 +14,6 @@ import (
 	"github.com/devspace-cloud/devspace/pkg/devspace/config/constants"
 	"github.com/devspace-cloud/devspace/pkg/devspace/config/generated"
 	"github.com/devspace-cloud/devspace/pkg/devspace/config/versions/latest"
-	"github.com/devspace-cloud/devspace/pkg/devspace/deploy/deployer/helm/merge"
-	"github.com/devspace-cloud/devspace/pkg/util/yamlutil"
 )
 
 // ConfigLoader is the base interface for the main config loader
@@ -85,19 +82,18 @@ func (l *configLoader) LoadGeneratedFromPath(path string) (*generated.Config, er
 
 // Exists checks whether the yaml file for the config exists or the configs.yaml exists
 func (l *configLoader) Exists() bool {
-	return configExistsInPath(".")
+	path := constants.DefaultConfigPath
+	if l.options.ConfigPath != "" {
+		path = l.options.ConfigPath
+	}
+
+	return configExistsInPath(path)
 }
 
 // configExistsInPath checks wheter a devspace configuration exists at a certain path
 func configExistsInPath(path string) bool {
 	// Check devspace.yaml
-	_, err := os.Stat(filepath.Join(path, constants.DefaultConfigPath))
-	if err == nil {
-		return true
-	}
-
-	// Check devspace-configs.yaml
-	_, err = os.Stat(filepath.Join(path, constants.DefaultConfigsPath))
+	_, err := os.Stat(path)
 	if err == nil {
 		return true
 	}
@@ -114,6 +110,7 @@ func (l *configLoader) New() *latest.Config {
 type ConfigOptions struct {
 	Profile     string
 	KubeContext string
+	ConfigPath  string
 
 	LoadedVars map[string]string
 	Vars       []string
@@ -203,8 +200,14 @@ func (l *configLoader) loadInternal(allowProfile bool) (*latest.Config, error) {
 		l.options.Profile = ""
 	}
 
+	// What path should we use
+	path := constants.DefaultConfigPath
+	if l.options.ConfigPath != "" {
+		path = l.options.ConfigPath
+	}
+
 	// Load base config
-	config, err := l.LoadFromPath(generatedConfig, constants.DefaultConfigPath)
+	config, err := l.LoadFromPath(generatedConfig, path)
 	if err != nil {
 		return nil, err
 	}
@@ -218,131 +221,12 @@ func (l *configLoader) loadInternal(allowProfile bool) (*latest.Config, error) {
 	return config, nil
 }
 
-func validate(config *latest.Config) error {
-	if config.Dev != nil {
-		if config.Dev.Ports != nil {
-			for index, port := range config.Dev.Ports {
-				if port.ImageName == "" && port.LabelSelector == nil {
-					return errors.Errorf("Error in config: imageName and label selector are nil in port config at index %d", index)
-				}
-				if port.PortMappings == nil {
-					return errors.Errorf("Error in config: portMappings is empty in port config at index %d", index)
-				}
-			}
-		}
-
-		if config.Dev.Sync != nil {
-			for index, sync := range config.Dev.Sync {
-				if sync.ImageName == "" && sync.LabelSelector == nil {
-					return errors.Errorf("Error in config: imageName and label selector are nil in sync config at index %d", index)
-				}
-			}
-		}
-
-		if config.Dev.Interactive != nil {
-			for index, imageConf := range config.Dev.Interactive.Images {
-				if imageConf.Name == "" {
-					return errors.Errorf("Error in config: Unnamed interactive image config at index %d", index)
-				}
-			}
-		}
-	}
-
-	if config.Commands != nil {
-		for index, command := range config.Commands {
-			if command.Name == "" {
-				return errors.Errorf("commands[%d].name is required", index)
-			}
-			if command.Command == "" {
-				return errors.Errorf("commands[%d].command is required", index)
-			}
-		}
-	}
-
-	if config.Hooks != nil {
-		for index, hookConfig := range config.Hooks {
-			if hookConfig.Command == "" {
-				return errors.Errorf("hooks[%d].command is required", index)
-			}
-		}
-	}
-
-	if config.Images != nil {
-		for imageConfigName, imageConf := range config.Images {
-			if imageConfigName == "" {
-				return errors.Errorf("images keys cannot be an empty string")
-			}
-			if imageConf == nil {
-				return errors.Errorf("images.%s is empty and should at least contain an image name", imageConfigName)
-			}
-			if imageConf.Image == "" {
-				return errors.Errorf("images.%s.image is required", imageConfigName)
-			}
-			if imageConf.Build != nil && imageConf.Build.Custom != nil && imageConf.Build.Custom.Command == "" {
-				return errors.Errorf("images.%s.build.custom.command is required", imageConfigName)
-			}
-			if imageConf.Image == "" {
-				return fmt.Errorf("images.%s.image is required", imageConfigName)
-			}
-		}
-	}
-
-	if config.Deployments != nil {
-		for index, deployConfig := range config.Deployments {
-			if deployConfig.Name == "" {
-				return errors.Errorf("deployments[%d].name is required", index)
-			}
-			if deployConfig.Helm == nil && deployConfig.Kubectl == nil {
-				return errors.Errorf("Please specify either helm or kubectl as deployment type in deployment %s", deployConfig.Name)
-			}
-			if deployConfig.Helm != nil && (deployConfig.Helm.Chart == nil || deployConfig.Helm.Chart.Name == "") && (deployConfig.Helm.ComponentChart == nil || *deployConfig.Helm.ComponentChart == false) {
-				return errors.Errorf("deployments[%d].helm.chart and deployments[%d].helm.chart.name or deployments[%d].helm.componentChart is required", index, index, index)
-			}
-			if deployConfig.Kubectl != nil && deployConfig.Kubectl.Manifests == nil {
-				return errors.Errorf("deployments[%d].kubectl.manifests is required", index)
-			}
-			if deployConfig.Helm != nil && deployConfig.Helm.ComponentChart != nil && *deployConfig.Helm.ComponentChart == true {
-				// Load override values from path
-				overwriteValues := map[interface{}]interface{}{}
-				if deployConfig.Helm.ValuesFiles != nil {
-					for _, overridePath := range deployConfig.Helm.ValuesFiles {
-						overwriteValuesPath, err := filepath.Abs(overridePath)
-						if err != nil {
-							return errors.Errorf("deployments[%d].helm.valuesFiles: Error retrieving absolute path from %s: %v", index, overridePath, err)
-						}
-
-						overwriteValuesFromPath := map[interface{}]interface{}{}
-						err = yamlutil.ReadYamlFromFile(overwriteValuesPath, overwriteValuesFromPath)
-						if err == nil {
-							merge.Values(overwriteValues).MergeInto(overwriteValuesFromPath)
-						}
-					}
-				}
-
-				// Load override values from data and merge them
-				if deployConfig.Helm.Values != nil {
-					merge.Values(overwriteValues).MergeInto(deployConfig.Helm.Values)
-				}
-
-				bytes, err := yaml.Marshal(overwriteValues)
-				if err != nil {
-					return errors.Errorf("deployments[%d].helm: Error marshaling overwrite values: %v", index, err)
-				}
-
-				componentValues := &latest.ComponentConfig{}
-				err = yaml.UnmarshalStrict(bytes, componentValues)
-				if err != nil {
-					return errors.Errorf("deployments[%d].helm.componentChart: component values are incorrect: %v", index, err)
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
 // SetDevSpaceRoot checks the current directory and all parent directories for a .devspace folder with a config and sets the current working directory accordingly
 func (l *configLoader) SetDevSpaceRoot() (bool, error) {
+	if l.options.ConfigPath != "" {
+		return configExistsInPath(l.options.ConfigPath), nil
+	}
+
 	cwd, err := os.Getwd()
 	if err != nil {
 		return false, err
@@ -357,7 +241,7 @@ func (l *configLoader) SetDevSpaceRoot() (bool, error) {
 	lastLength := 0
 	for len(cwd) != lastLength {
 		if cwd != homedir {
-			configExists := configExistsInPath(cwd)
+			configExists := configExistsInPath(filepath.Join(cwd, constants.DefaultConfigPath))
 			if configExists {
 				// Change working directory
 				err = os.Chdir(cwd)
